pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "browser-agent-bridge"
version = "0.1.0"
description = "Browser Bridge server and CLI for controlling a Chrome extension over WebSocket"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
keywords = ["browser", "bridge", "automation", "websocket", "fastapi", "cli"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Framework :: FastAPI",
  "Topic :: Internet :: WWW/HTTP :: WSGI :: Server",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "fastapi>=0.116.1",
  "httpx>=0.28.1",
  "PyJWT>=2.10.1",
  "uvicorn[standard]>=0.35.0",
]

[project.optional-dependencies]
dev = [
  "build>=1.2.2",
  "pytest>=8.4.2",
  "twine>=5.1.1",
]

[project.scripts]
browser-bridge = "browser_bridge.cli:main"
browser-bridge-server = "browser_bridge.server:main"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.setuptools]
packages = ["browser_bridge"]
